+++ oop/Hierarchy/developer.py
@@ -1,0 +1,3 @@
+from .employee import Employee
+class Developer(Employee):
+    pass
